Fixed weekly recurrences with an interval > 1
The previous implementation decided whether the weekly rule was valid based on the commercial week in the year. i.e. for @interval = 2, the rule would be valid on the 2nd, 4th, 6th etc week of the year, regardless of the start date of the schedule.

This commit changes it such that the rule is valid on the week of the start date of the schedule, and then every @interval weeks after that. 

This is achieved by finding the difference in weeks between the start date and the tested date and validating it if it is a multiple of @interval.

Added a spec and updated the existing interval=2 spec to actually set interval to 2.