Made closing_time argument optional in occurrences
End time is optional for `occurrences` method. For instance if for a terminating schedule I call `schedule.occurrences`, it throws `wrong number of arguments (0 for 1)` which is ambiguous. It works if I pass `nil`. Since anyway, there is no check to prevent infinite looping incase of non-terminating schedules, I propose that this argument should be made optional.