Add deprecations and settings for compatibility
Changing :start_date to :start_time everywhere will take a while for
some users due to serialized schedules that still have this key.

For now we'll export both start_time and start_date in serialized
output but this duplication should be removed in the next version.

We will continue to read both :start_date and :start_time, with a
preference for the new :start_time.