Avoid serializing YAML for blank values
When dumping to the database, ActiveRecord would run blank values (nil
and "") through the serializer and wrap it a YAML string that is not
valid for us to parse as a schedule. When the attribute is blank going
into the serialized format, it should remain blank coming back out.

Fixed #231