Realign first wday validation to the selected week_start
When using a non-zero (non-Sunday) week start, the rule validations need to
choose the first wday to align onto when jumping into the schedule. The first
wday needs to be based on the correct week start, since 0 (Sunday) then comes at
the end of the week and causes the next valid occurrence to be skipped.